stm_serial.py, sonar_filter.py : imu 연결 끊겼을때 재연결 srv 실행

+++ zetabank_robot_control/zetabank_bringup/scripts/stm_serial.py
@@ -4,6 +4,8 @@
 import rospy
 import os,sys,threading, time
 from sensor_msgs.msg import Imu
+from std_msgs.msg import String
+from std_srvs.srv import Empty, EmptyResponse, EmptyRequest
 
 class Ros_check:
     def __init__(self, imu_topic):
@@ -13,31 +15,66 @@ class Ros_check:
         self.check_flag = True
 
 class Command:
-    def __init__(self):
+    
+    def __init__(self):    
         self.command_rosrun = "rosrun rosserial_python serial_node.py _port:=/dev/ttyACM{} _baud:=115200 __name:='STM'"
-    def ACM_check(self):
-        for i in range(2):
-            command_rosrun = self.command_rosrun.format(i)
-            t1 = threading.Thread(target= ACM_thread, name= "ACM thread", args=(command_rosrun,))
-            t1.setDaemon(True)
-            t1.start()
-            #ros
-            imu_topic = '/imu'
-            ros_check = Ros_check(imu_topic)
-            time.sleep(10)
+        self.number = 99
+        
+        self.imu_error_service = rospy.Service('/emergency_stm', Empty, self.ACM_check)
 
-            if not ros_check.check_flag:
-                nodes = os.popen("rosnode lsit").readline()
-                os.system("rosnode kill /STM")
-                t1.join()
+    def ACM_check(self,msg = EmptyRequest()):
+        time.sleep(0.2)
+        os.system('rosnode kill /STM')
+        while True:          
+            lsusb = self.STM_usb()
+            if lsusb:
+                print('STM_on')
+                if self.number != 99: 
+                    time.sleep(2)
+                    self.ACM_double_check(self.number)
+                    return EmptyResponse()
+
+                else :
+                    for i in range(2):        
+                        t1 = self.ACM_double_check(i)
+                    #ros
+                        imu_topic = '/imu'
+                        ros_check = Ros_check(imu_topic)
+                        time.sleep(10)
+
+                        if not ros_check.check_flag:
+                            nodes = os.popen("rosnode list").readline()
+                            os.system("rosnode kill /STM")
+                            t1.join()
+                        else:
+                            self.number = i       
+                            break  
+                break  
             else:
-                break
-                
+                pass
+                # return EmptyResponse()
 
-def ACM_thread(command_rosrun):
-    os.system(command_rosrun)
+    def ACM_double_check(self,number): 
+        command_rosrun = self.command_rosrun.format(number)
+        t1 = threading.Thread(target= self.ACM_thread, name="ACM thread", args=(command_rosrun,))
+        t1.setDaemon(True)
+        t1.start()
+        return t1   
 
+    def STM_usb(self):
+        lsusb_list = os.popen("lsusb").read()
+        if 'STM' in lsusb_list:
+            return True
+        else:
+            return False
+
+    def ACM_thread(self,command_rosrun):
+        os.system(command_rosrun)
 
 if __name__=='__main__':
     rospy.init_node("stm_starter")
-    Command().ACM_check()
+    acm_command = Command()
+
+    acm_command.ACM_check()
+
+    rospy.spin()
